correção de alguns codigos + adição do login no redis

+++ redis/compras.py
@@ -8,7 +8,7 @@ redis_client = redis.Redis(
     password=REDIS_CONFIG["password"]
 )
 
-def create_compra():
+def create_compra(cpf):
     print("\nInserir nova compra")
     cpf_usuario = input("CPF do usuário: ")
     nome_produto = input("Nome do produto: ")
@@ -26,7 +26,7 @@ def create_compra():
 
     print("Compra registrada com sucesso.")
 
-def read_compras():
+def read_compras(cpf):
     cpf_usuario = input("Digite o CPF do usuário: ")
 
     compras_data = redis_client.lrange(f"compras:{cpf_usuario}", 0, -1)
@@ -39,32 +39,7 @@ def read_compras():
     else:
         print("Nenhuma compra encontrada.")
 
-def update_compra():
-    cpf_usuario = input("Digite o CPF do usuário: ")
-    nome_produto = input("Digite o nome do produto da compra que deseja atualizar: ")
-
-    compra_key = f"compras:{cpf_usuario}"
-    compras_data = redis_client.lrange(compra_key, 0, -1)
-
-    updated_compras = []
-
-    for compra_json in compras_data:
-        compra_data = json.loads(compra_json)
-
-        if compra_data["nome_produto"] == nome_produto:
-            print(f"Atualizando compra: {json.dumps(compra_data, indent=4)}")
-
-            compra_data["quantidade"] = input("Nova quantidade: ")
-
-            updated_compras.append(json.dumps(compra_data))
-            print("Compra atualizada com sucesso.")
-        else:
-            updated_compras.append(compra_json)
-
-    redis_client.delete(compra_key)
-    redis_client.rpush(compra_key, *updated_compras)
-
-def delete_compra():
+def delete_compra(cpf):
     cpf_usuario = input("Digite o CPF do usuário: ")
     nome_produto = input("Digite o nome do produto da compra que deseja excluir: ")
 
@@ -81,16 +56,18 @@ def delete_compra():
         else:
             deleted_compras.append(compra_json)
 
-    redis_client.delete(compra_key)
-    redis_client.rpush(compra_key, *deleted_compras)
+    if deleted_compras:
+        redis_client.delete(compra_key)
+        redis_client.rpush(compra_key, *deleted_compras)
+    else:
+        print("Nenhuma compra encontrada para remoção.")
 
 def menu_compras(cpf):
     while True:
         print("\n\033[1mMenu Compras:\033[0m")
         print("1 - Criar Compra")
         print("2 - Ler Compras")
-        print("3 - Atualizar Compra")
-        print("4 - Deletar Compra")
+        print("3 - Deletar Compra")
         print("\nV - Voltar")
 
         sub = input("Digite a opção desejada: ").upper()
@@ -105,7 +82,4 @@ def menu_compras(cpf):
             read_compras(cpf)
 
         elif sub == '3':
-            update_compra(cpf)
-
-        elif sub == '4':
             delete_compra(cpf)
